# [4.4] [Core] Fix performance issues in some `CowData` methods