These relied on `get` which does a bounds check each time on debug builds

In my testing, using the below MRP, this gives (before vs after):
```
PackedInt32Array.find():  607259 usec for 5000000 iterations
PackedInt32Array.rfind(): 483297 usec for 5000000 iterations
PackedInt32Array.count(): 747576 usec for 5000000 iterations
```
```
PackedInt32Array.find():  480270 usec for 5000000 iterations
PackedInt32Array.rfind(): 465183 usec for 5000000 iterations
PackedInt32Array.count(): 470655 usec for 5000000 iterations
```

[test-cowdata.zip](https://github.com/user-attachments/files/18550508/test-cowdata.zip)

There might be more points of issue here (`CowData(std::initializer_list<T> p_init)` calling `set` repeatedly and causing COW for example) but these are the most trivial cases

<!--
Please target the `master` branch in priority.

Relevant fixes are cherry-picked for stable branches as needed by maintainers.

To speed up the contribution process and avoid CI errors, please set up pre-commit hooks locally:
https://docs.godotengine.org/en/latest/contributing/development/code_style_guidelines.html
-->