@AThousandShips 2025-01-26
Will add some further parts to the test project above for testing cases where `find` and `rfind` don't find any element as well